+++ doc/authoring_command_modules/example_module_template/azure/cli/command_modules/example/_help.py
@@ -8,6 +8,6 @@ from knack.help_files import helps
 # pylint: disable=line-too-long
 
 helps['example'] = """
-            type: group
-            short-summary: Example group description.
-            """
+    type: group
+    short-summary: Example group description.
+"""