+++ src/command_modules/azure-cli-backup/azure/cli/command_modules/backup/tests/test_backup_commands.py
@@ -57,6 +57,7 @@ class BackupTests(ScenarioTest, unittest.TestCase):
         restore_cmd_string += ' -g {} -v {}'.format(resource_group, vault_name)
         restore_cmd_string += ' -c {} -i {} -r {}'.format(container, item, recovery_point)
         restore_cmd_string += ' --storage-account {} --query name'.format(storage_account)
+        restore_cmd_string += ' --restore-to-staging-storage-account'
         trigger_restore_job_name = self.cmd(restore_cmd_string).get_output_in_json()
         self.cmd('backup job wait -g {} -v {} -n {}'.format(resource_group, vault_name, trigger_restore_job_name))
 
@@ -235,6 +236,12 @@ class BackupTests(ScenarioTest, unittest.TestCase):
                      JMESPathCheck("length(@)", 1),
                      JMESPathCheck("length([?properties.friendlyName == '{}'])".format(vm2), 1)])
 
+        self.cmd('backup item list -g {} -v {}'
+                 .format(resource_group, vault_name), checks=[
+                     JMESPathCheck("length(@)", 2),
+                     JMESPathCheck("length([?properties.friendlyName == '{}'])".format(vm1), 1),
+                     JMESPathCheck("length([?properties.friendlyName == '{}'])".format(vm2), 1)])
+
         self.cmd('backup item set-policy -g {} -v {} -c {} -n {} -p {}'
                  .format(resource_group, vault_name, container1, vm1, policy_name), checks=[
                      JMESPathCheck("properties.entityFriendlyName", vm1),
@@ -318,11 +325,20 @@ class BackupTests(ScenarioTest, unittest.TestCase):
         rp_name = self.cmd('backup recoverypoint list -g {} -v {} -c {} -i {} --query [0].name'
                            .format(resource_group, vault_name, vm_name, vm_name)).get_output_in_json()
 
+        # Original Storage Account Restore Fails
+        osa_restore_cmd_string = 'backup restore restore-disks'
+        osa_restore_cmd_string += ' -g {} -v {}'.format(resource_group, vault_name)
+        osa_restore_cmd_string += ' -c {} -i {} -r {}'.format(vm_name, vm_name, rp_name)
+        osa_restore_cmd_string += ' --storage-account {}'.format(storage_account)
+        osa_restore_cmd_string += ' --restore-to-staging-storage-account false'
+        self.cmd(osa_restore_cmd_string, expect_failure=True)
+
         # Trigger Restore
         restore_cmd_string = 'backup restore restore-disks'
         restore_cmd_string += ' -g {} -v {}'.format(resource_group, vault_name)
         restore_cmd_string += ' -c {} -i {} -r {}'.format(vm_name, vm_name, rp_name)
         restore_cmd_string += ' --storage-account {}'.format(storage_account)
+        restore_cmd_string += ' --restore-to-staging-storage-account'
         trigger_restore_job_json = self.cmd(restore_cmd_string, checks=[
             JMESPathCheck("properties.entityFriendlyName", vm_name),
             JMESPathCheck("properties.operation", "Restore"),
@@ -364,6 +380,7 @@ class BackupTests(ScenarioTest, unittest.TestCase):
         restore_cmd_string += ' -g {} -v {}'.format(resource_group, vault_name)
         restore_cmd_string += ' -c {} -i {} -r {}'.format(vm_name, vm_name, rp_name)
         restore_cmd_string += ' --storage-account {} --query name'.format(storage_account)
+        restore_cmd_string += ' --restore-to-staging-storage-account'
         trigger_restore_job_name = self.cmd(restore_cmd_string).get_output_in_json()
 
         self.cmd('backup job show -g {} -v {} -n {}'