+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -71,13 +71,12 @@ extension_mappings = {
 
 
 def _construct_identity_info(identity_scope, identity_role, port, external_identities):
-    info = {
-        'scope': identity_scope or '',
-        'role': str(identity_role),  # could be DefaultStr, so convert to string
-        'port': port
-    }
+    info = {'port': port}
+    if identity_scope:
+        info['scope'] = identity_scope
+        info['role'] = str(identity_role)  # could be DefaultStr, so convert to string
     if external_identities:
-        info['externalIdentities'] = external_identities
+        info['userAssignedIdentities'] = external_identities
     return info
 
 
@@ -1675,7 +1674,7 @@ def create_vmss(cmd, vmss_name, resource_group_name, image,
                 single_placement_group=None, custom_data=None, secrets=None,
                 plan_name=None, plan_product=None, plan_publisher=None, plan_promotion_code=None, license_type=None,
                 assign_identity=None, identity_scope=None, identity_role='Contributor',
-                identity_role_id=None, zones=None):
+                identity_role_id=None, zones=None, priority=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.command_modules.vm._template_builder import (ArmTemplateBuilder, StorageProfile, build_vmss_resource,
@@ -1683,7 +1682,7 @@ def create_vmss(cmd, vmss_name, resource_group_name, image,
                                                                 build_load_balancer_resource,
                                                                 build_vmss_storage_account_pool_resource,
                                                                 build_application_gateway_resource,
-                                                                build_msi_role_assignment)
+                                                                build_msi_role_assignment, build_nsg_resource)
     subscription_id = get_subscription_id(cmd.cli_ctx)
     network_id_template = resource_id(
         subscription=subscription_id, resource_group=resource_group_name,
@@ -1748,6 +1747,10 @@ def create_vmss(cmd, vmss_name, resource_group_name, image,
 
     # Handle load balancer creation
     if load_balancer_type == 'new':
+        # Defaults SKU to 'Standard' for zonal scale set
+        if load_balancer_sku is None:
+            load_balancer_sku = 'Standard' if zones else 'Basic'
+
         vmss_dependencies.append('Microsoft.Network/loadBalancers/{}'.format(load_balancer))
 
         lb_dependencies = []
@@ -1851,6 +1854,14 @@ def create_vmss(cmd, vmss_name, resource_group_name, image,
     if secrets:
         secrets = _merge_secrets([validate_file_or_dict(secret) for secret in secrets])
 
+    if nsg is None and zones:
+        # Per https://docs.microsoft.com/en-us/azure/load-balancer/load-balancer-standard-overview#nsg
+        nsg_name = '{}NSG'.format(vmss_name)
+        master_template.add_resource(build_nsg_resource(
+            None, nsg_name, location, tags, 'rdp' if os_type.lower() == 'windows' else 'ssh'))
+        nsg = "[resourceId('Microsoft.Network/networkSecurityGroups', '{}')]".format(nsg_name)
+        vmss_dependencies.append('Microsoft.Network/networkSecurityGroups/{}'.format(nsg_name))
+
     vmss_resource = build_vmss_resource(cmd, vmss_name, naming_prefix, location, tags,
                                         not disable_overprovision, upgrade_policy_mode,
                                         vm_sku, instance_count,
@@ -1865,7 +1876,7 @@ def create_vmss(cmd, vmss_name, resource_group_name, image,
                                         backend_address_pool_id, inbound_nat_pool_id, health_probe=health_probe,
                                         single_placement_group=single_placement_group,
                                         custom_data=custom_data, secrets=secrets,
-                                        license_type=license_type, zones=zones)
+                                        license_type=license_type, zones=zones, priority=priority)
     vmss_resource['dependsOn'] = vmss_dependencies
 
     if plan_name: