+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/tests/test_vm_commands.py
@@ -1748,6 +1748,7 @@ class VMSSLoadBalancerWithSku(ScenarioTest):
     def test_vmss_lb_sku(self, resource_group):
 
         self.kwargs.update({
+            'vmss0': 'vmss0',
             'vmss': 'vmss1',
             'lb': 'lb1',
             'ip': 'pubip1',
@@ -1755,6 +1756,15 @@ class VMSSLoadBalancerWithSku(ScenarioTest):
             'loc': 'eastus2'
         })
 
+        # default to Basic
+        self.cmd('vmss create -g {rg} -l {loc} -n {vmss0} --image UbuntuLTS --admin-username admin123 --admin-password PasswordPassword1!')
+        self.cmd('network lb list -g {rg}', checks=self.check('[0].sku.name', 'Basic'))
+        self.cmd('network public-ip list -g {rg}', checks=[
+            self.check('[0].sku.name', 'Basic'),
+            self.check('[0].publicIpAllocationMethod', 'Dynamic')
+        ])
+
+        # but you can overrides the defaults
         self.cmd('vmss create -g {rg} -l {loc} -n {vmss} --lb {lb} --lb-sku {sku} --public-ip-address {ip} --image UbuntuLTS --admin-username admin123 --admin-password PasswordPassword1!')
         self.cmd('network lb show -g {rg} -n {lb}',
                  checks=self.check('sku.name', 'Standard'))
@@ -1897,7 +1907,8 @@ class MSIScenarioTest(ScenarioTest):
 
         # create a linux vm with identity but w/o a role assignment (--scope "")
         self.cmd('vm create -g {rg} -n {vm1} --image debian --assign-identity --admin-username admin123 --admin-password PasswordPassword1!', checks=[
-            self.check('identity.scope', ''),
+            self.check('identity.scope', None),
+            self.check('identity.role', None),
             self.check('identity.port', 50342)
         ])
         # the extension should still get provisioned
@@ -1908,7 +1919,7 @@ class MSIScenarioTest(ScenarioTest):
 
         # create a vmss with identity but w/o a role assignment (--scope "")
         self.cmd('vmss create -g {rg} -n {vmss1} --image debian --assign-identity --admin-username admin123 --admin-password PasswordPassword1!', checks=[
-            self.check('vmss.identity.scope', ''),
+            self.check('vmss.identity.scope', None),
             self.check('vmss.identity.port', 50342)
         ])
 
@@ -1922,7 +1933,7 @@ class MSIScenarioTest(ScenarioTest):
         self.cmd('vm create -g {rg} -n {vm2} --image debian --admin-username admin123 --admin-password PasswordPassword1!')
         # assign identity but w/o a role assignment
         self.cmd('vm assign-identity -g {rg} -n {vm2}', checks=[
-            self.check('scope', ''),
+            self.check('scope', None),
             self.check('port', 50342)
         ])
         # the extension should still get provisioned
@@ -1935,7 +1946,7 @@ class MSIScenarioTest(ScenarioTest):
         # skip playing back till the test issue gets addressed https://github.com/Azure/azure-cli/issues/4016
         if self.is_live:
             self.cmd('vmss assign-identity -g {rg} -n {vmss2}', checks=[
-                self.check('scope', ''),
+                self.check('scope', None),
                 self.check('port', 50342)
             ])
 
@@ -1960,9 +1971,17 @@ class MSIScenarioTest(ScenarioTest):
                                checks=self.check('name', '{emsi}')).get_output_in_json()
         emsi2_result = self.cmd('identity create -g {rg} -n {emsi2}').get_output_in_json()
 
+        # create a vm with only user assigned identity
+        result = self.cmd('vm create -g {rg} -n vm2 --image ubuntults --assign-identity {emsi} --generate-ssh-keys', checks=[
+            self.check('identity.role', None),
+            self.check('identity.scope', None),
+            self.check('length(identity.userAssignedIdentities)', 1)
+        ]).get_output_in_json()
+        self.assertEqual(result['identity']['userAssignedIdentities'][0].lower(), emsi_result['id'].lower())
+
         # create a vm with system + user assigned identities
         result = self.cmd('vm create -g {rg} -n {vm} --image ubuntults --assign-identity {emsi} [system] --role reader --scope {scope} --generate-ssh-keys --admin-username ubuntuadmin').get_output_in_json()
-        self.assertEqual(result['identity']['externalIdentities'][0].lower(), emsi_result['id'].lower())
+        self.assertEqual(result['identity']['userAssignedIdentities'][0].lower(), emsi_result['id'].lower())
         result = self.cmd('vm show -g {rg} -n {vm}', checks=[
             self.check('length(identity.identityIds)', 1),
             self.check('identity.type', 'SystemAssigned, UserAssigned')
@@ -2004,7 +2023,7 @@ class MSIScenarioTest(ScenarioTest):
 
         # create a vmss with system + user assigned identities
         result = self.cmd('vmss create -g {rg} -n {vmss} --image ubuntults --assign-identity {emsi} [system] --role reader --scope {scope} --instance-count 1 --generate-ssh-keys --admin-username ubuntuadmin').get_output_in_json()
-        self.assertEqual(result['vmss']['identity']['externalIdentities'][0].lower(), emsi_result['id'].lower())
+        self.assertEqual(result['vmss']['identity']['userAssignedIdentities'][0].lower(), emsi_result['id'].lower())
 
         result = self.cmd('vmss show -g {rg} -n {vmss}', checks=[
             self.check('length(identity.identityIds)', 1),
@@ -2078,7 +2097,7 @@ class VMZoneScenarioTest(ScenarioTest):
         self.assertTrue(set([resource_group_location, self.kwargs['zones']]).issubset(table_output))
 
     @ResourceGroupPreparer(name_prefix='cli_test_vmss_zones', location='eastus2')
-    def test_vmss_create_zones(self, resource_group, resource_group_location):
+    def test_vmss_create_single_zone(self, resource_group, resource_group_location):
 
         self.kwargs.update({
             'zones': '2',
@@ -2093,6 +2112,36 @@ class VMZoneScenarioTest(ScenarioTest):
         result = self.cmd('vmss list -g {rg} -otable')
         table_output = set(result.output.splitlines()[2].split())
         self.assertTrue(set([resource_group_location, self.kwargs['vmss'], self.kwargs['zones']]).issubset(table_output))
+
+        self.cmd('network lb list -g {rg}', checks=[
+            self.check('[0].sku.name', 'Standard')
+        ])
+        self.cmd('network public-ip list -g {rg}', checks=[
+            self.check('[0].sku.name', 'Standard'),
+            self.check('[0].zones', ['2'])
+        ])
+
+    @ResourceGroupPreparer(name_prefix='cli_test_vmss_zones', location='eastus2')
+    def test_vmss_create_x_zones(self, resource_group, resource_group_location):
+
+        self.kwargs.update({
+            'zones': '1 2 3',
+            'vmss': 'vmss123'
+        })
+        self.cmd('vmss create -g {rg} -n {vmss} --admin-username clitester --admin-password PasswordPassword1! --image debian --zones {zones}')
+        self.cmd('vmss show -g {rg} -n {vmss}',
+                 checks=self.check('zones', ['1', '2', '3']))
+        result = self.cmd('vmss show -g {rg} -n {vmss} -otable')
+        table_output = set(result.output.splitlines()[2].split())
+        self.assertTrue(set([resource_group_location, self.kwargs['vmss']] + self.kwargs['zones'].split()).issubset(table_output))
+
+        self.cmd('network lb list -g {rg}', checks=[
+            self.check('[0].sku.name', 'Standard')
+        ])
+        self.cmd('network public-ip list -g {rg}', checks=[
+            self.check('[0].sku.name', 'Standard'),
+            self.check('[0].zones', None)
+        ])
 
     @ResourceGroupPreparer(name_prefix='cli_test_disk_zones', location='eastus2')
     def test_disk_create_zones(self, resource_group, resource_group_location):
@@ -2222,6 +2271,17 @@ class VMSSRollingUpgrade(ScenarioTest):
 
         # 'cancel' should fail as we have no active upgrade to cancel
         self.cmd('vmss rolling-upgrade cancel -g {rg} -n {vmss}', expect_failure=True)
+
+
+class VMSSPriorityTesting(ScenarioTest):
+    @ResourceGroupPreparer(name_prefix='vmss_low_pri', location='CentralUSEUAP')
+    def test_vmss_create_with_low_priority(self, resource_group, resource_group_location):
+        self.kwargs.update({
+            'priority': 'Low',
+            'vmss': 'vmss123'
+        })
+        self.cmd('vmss create -g {rg} -n {vmss} --admin-username clitester --admin-password PasswordPassword1! --image debian --priority {priority}')
+        self.cmd('vmss show -g {rg} -n {vmss}', checks=self.check('virtualMachineProfile.priority', self.kwargs['priority']))
 
 
 class VMLBIntegrationTesting(ScenarioTest):