+++ src/command_modules/azure-cli-servicefabric/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.0.9
+++++++
+* Added detailed errors to validation response when creating cluster.
+* Fix missing client issue with several commands.
+
 0.0.8
 ++++++
 * Update for CLI core changes.